+++ com.unity.render-pipelines.high-definition/Editor/Wizard/HDWizard.UIElement.cs
@@ -1,5 +1,7 @@
 using UnityEngine;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using UnityEngine.Rendering;
@@ -157,23 +159,6 @@ namespace UnityEditor.Rendering.HighDefinition
             }
         }
 
-        void Repopulate()
-        {
-            if (!AssetDatabase.IsValidFolder("Assets/" + HDProjectSettings.projectSettingsFolderPath))
-                AssetDatabase.CreateFolder("Assets", HDProjectSettings.projectSettingsFolderPath);
-
-            var hdrpAsset = ScriptableObject.CreateInstance<HDRenderPipelineAsset>();
-            hdrpAsset.name = "HDRenderPipelineAsset";
-
-            AssetDatabase.CreateAsset(hdrpAsset, "Assets/" + HDProjectSettings.projectSettingsFolderPath + "/" + hdrpAsset.name + ".asset");
-
-            GraphicsSettings.renderPipelineAsset = hdrpAsset;
-            if (!IsHdrpAssetRuntimeResourcesCorrect())
-                FixHdrpAssetRuntimeResources(true);
-            if (!IsHdrpAssetEditorResourcesCorrect())
-                FixHdrpAssetEditorResources(true);
-        }
-
         #endregion
 
         #region UIELEMENT
@@ -325,7 +310,7 @@ namespace UnityEditor.Rendering.HighDefinition
                 base.CheckUpdate();
                 if (currentStatus)
                 {
-                    foreach (VisualElementUpdatable updatable in Children())
+                    foreach (VisualElementUpdatable updatable in Children().Where(e => e is VisualElementUpdatable))
                         updatable.CheckUpdate();
                 }
             }
@@ -411,7 +396,7 @@ namespace UnityEditor.Rendering.HighDefinition
                         this.Q(name: "StatusError").style.display = DisplayStyle.None;
                     }
                     this.Q(name: "Resolver").style.display = DisplayStyle.None;
-                    this.Q(name: "HelpBox").style.display = DisplayStyle.None;
+                    this.Q(className: "HelpBox").style.display = DisplayStyle.None;
                 }
                 else
                 {
@@ -421,7 +406,7 @@ namespace UnityEditor.Rendering.HighDefinition
                         this.Q(name: "StatusError").style.display = !statusOK ? (m_SkipErrorIcon ? DisplayStyle.None : DisplayStyle.Flex) : DisplayStyle.None;
                     }
                     this.Q(name: "Resolver").style.display = statusOK || !haveFixer ? DisplayStyle.None : DisplayStyle.Flex;
-                    this.Q(name: "HelpBox").style.display = statusOK ? DisplayStyle.None : DisplayStyle.Flex;
+                    this.Q(className: "HelpBox").style.display = statusOK ? DisplayStyle.None : DisplayStyle.Flex;
                 }
             }
         }
@@ -483,7 +468,7 @@ namespace UnityEditor.Rendering.HighDefinition
                 this.label = new Label(message);
                 icon = new Image();
 
-                name = "HelpBox";
+                AddToClassList("HelpBox");
                 Add(icon);
                 Add(this.label);
 
@@ -507,6 +492,29 @@ namespace UnityEditor.Rendering.HighDefinition
 
             protected override void UpdateDisplay(bool statusOK, bool haveFixer)
                 => this.Q(name: "FixAll").style.display = statusOK ? DisplayStyle.None : DisplayStyle.Flex;
+        }
+
+        class ScopeBox : VisualElementUpdatable
+        {
+            readonly Label label;
+            bool initTitleBackground;
+
+            public ScopeBox(string title) : base(null, false)
+            {
+                label = new Label(title);
+                label.name = "Title";
+                AddToClassList("ScopeBox");
+                Add(label);
+            }
+
+            public override void CheckUpdate()
+            {
+                foreach (VisualElementUpdatable updatable in Children().Where(e => e is VisualElementUpdatable))
+                    updatable.CheckUpdate();
+            }
+
+            protected override void UpdateDisplay(bool statusOK, bool haveFixer)
+            {}
         }
 
         #endregion