+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [11.0.1]
+
+### Fixed
+- Fixed missing Update in Wizard's DXR Documentation
+
+### Changed
+- Updated Wizard to better handle RenderPipelineAsset in Quality Settings
+
 ## [11.0.0] - 2020-12-02
 
 ### Added